requirements.txt
PyMuPDF==1.23.8
streamlit==1.31.0
google-generativeai==0.4.1
faiss-cpu==1.7.4
python-dotenv==1.0.1
Pillow==10.2.0
langchain==0.1.9
langchain-google-genai==0.0.9
numpy==1.26.4
reportlab==4.1.0
